pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "openlimno"
version = "0.1.0-dev"
description = "Open-source water ecology modeling platform — PHABSIM/IFIM modernization with SCHISM 2D wrapping"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { file = "LICENSE" }
authors = [
    { name = "OpenLimno Contributors", email = "hello@openlimno.org" },
]
keywords = [
    "ecology", "hydrology", "fish-habitat", "PHABSIM", "IFIM",
    "ecological-flow", "WUA", "HSI", "river2d", "SCHISM",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Hydrology",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "xarray>=2024.1",
    "netCDF4>=1.6",
    "pyarrow>=15",
    "pyproj>=3.6",
    "shapely>=2.0",
    "geopandas>=0.14",
    "rasterio>=1.3",
    "matplotlib>=3.8",
    "pyyaml>=6",
    "jsonschema>=4.21",
    "click>=8.1",
    "rich>=13",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-regressions>=2",
    "hypothesis>=6",
    "ruff>=0.4",
    "mypy>=1.10",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]
schism = [
    "pyschism",
]

[project.urls]
Homepage = "https://openlimno.org"
Documentation = "https://docs.openlimno.org"
Repository = "https://github.com/openlimno/openlimno"
Issues = "https://github.com/openlimno/openlimno/issues"
Changelog = "https://github.com/openlimno/openlimno/blob/main/CHANGELOG.md"

[project.scripts]
openlimno = "openlimno.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openlimno"]

[tool.hatch.build.targets.wheel.shared-data]
"src/openlimno/wedm/schemas" = "openlimno/wedm/schemas"

# ---------------------------------------------------------------
# ruff
# ---------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",          # pycodestyle / pyflakes
    "I",                      # isort
    "N",                      # naming
    "UP",                     # pyupgrade
    "B",                      # bugbear
    "SIM",                    # simplify
    "RUF",                    # ruff-specific
    "PT",                     # pytest style
]
ignore = [
    "E501",  # line too long (formatter handles)
    "N803",  # argument lowercase (math vars like Q, A, U_burst are intentional)
    "N806",  # var name in func (same)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N806", "B011"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------
# mypy
# ---------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
plugins = []

[[tool.mypy.overrides]]
module = "tests.*"
strict = false

# ---------------------------------------------------------------
# pytest
# ---------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests", "benchmarks"]
addopts = "-ra --strict-markers"
markers = [
    "slow: tests that take >5s",
    "schism: requires external SCHISM binary",
    "qgis: requires QGIS environment",
    "benchmark: validation benchmark, not unit test",
]

[tool.coverage.run]
source = ["src/openlimno"]
omit = ["src/openlimno/qgis/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
